reconstruct module

+++ sources/btc_math.move
@@ -12,7 +12,7 @@ public fun to_u32(v: vector<u8>): u32 {
     let mut ans = 0u32;
     let mut i = 0u8;
     while (i < 4) {
-        ans = ans + (v[i] as u32 << i*8);
+        ans = ans + ((v[i as u64] as u32) << i*8);
         i = i + 1;
     };
